synx: dma_fence_wait should mark APSS as waiter
Currently, when a client calls dma_fence_add_callback or
dma_fence_wait, Synx is unaware of it. As a result, APSS is not
marked as a waiter in global memory and hence does not receive
signal from other cores.

This change marks APSS as waiter when both these functions call
enable_signaling callback registered by Synx.

Change-Id: I16a147bcbe4ba6aa195bd5460beb80054a7246e8
Signed-off-by: Pravin Kumar Ravi <quic_pravinku@quicinc.com>
Signed-off-by: Amir Suhail <quic_asuhail@quicinc.com>