msm: synx: synx_import fix for incorrect dma flag check
In synx_import API call, if flags is passed as combination of SYNX and
DMA and desc.id[0] is set to zero but synx_obj is non zero, then it will
try to import the synx_obj value as dma fence and causes a crash in
dma_fence_get API in synx_util_init_coredata.
Dma_fence should always be imported after a valid sync_file_get_fence
check.

Change-Id: I57a46ba1bfc0024f53459fa27174dccfce1a9308
Signed-off-by: Amir Suhail <quic_asuhail@quicinc.com>