msm: synx: Fixed race  which incorrectly marked waiter
On apps in race condition we were storing parents of a child
locally and later using parents from local variable to set
waiter for that parent. If by the time we are setting parent
as waiter the parent handle is released by the other core
we will incorrectly set waiters variable in global memory
for parent even though that index is cleared. This fix does
not mark apps as waiter if the global entry for that handle
is no longer valid. The race condition was in
synx_global_test_status_set_parent_child_wait
where synx_global_set_waiting_core gets called.

Change-Id: I63f15715ec0e9475ca5dbb79f5a67c4752dae10b
Signed-off-by: Kuldeep Singh <quic_kulsin@quicinc.com>