msm: synx: Avoid returning already signaled status for dma-fence
There is a race condition if dma-fence is getting signaled and ipc lite
signaling is also received on APPS. The ipc_lite signal will try to
update the dma-fence. Before signaling the dma-fence from synx side we
remove the dma-fence callback and then again check if dma-fence was
signaled or not.
In parallel hw-fence is signaling the dma-fence. Since the dma callback
is removed by synx, we will not get synx_fence_callback and in ipc_lite
signal processing we will get already signal value. This race condition
causes a bug in which both of the signals doesn't dispatch the async
callbacks.

Change-Id: I43352d2017b8886493b560e8c1686a496313e3fb
Signed-off-by: Amir Suhail <quic_asuhail@quicinc.com>